pubsubSchemaCapableIOProvider subclasses rather than external classes
